+++ src/accessories/smarthome/button/button.accessory.js
@@ -9,6 +9,7 @@ class Accessory {
     this.accessory = accessory;
     this.accessories = accessories;
     this.meshMaster = meshMaster;
+    this.BatteryService = this.api.hap.Service.Battery || this.api.hap.Service.BatteryService;
 
     //Buttons needs own Handler due to shorter polling timer
     this.handler = Handler.configure(
@@ -37,14 +38,14 @@ class Accessory {
     const buttons = this.accessory.context.config.buttons;
 
     if (buttons && this.accessory.context.config.battery) {
-      let batteryService = this.accessory.getService(this.api.hap.Service.BatteryService);
+      let batteryService = this.accessory.getService(this.BatteryService);
 
       if (!batteryService) {
         logger.info(
           'Adding Battery service',
           `${this.accessory.displayName} (${this.accessory.context.config.subtype})`
         );
-        batteryService = this.accessory.addService(this.api.hap.Service.BatteryService);
+        batteryService = this.accessory.addService(this.BatteryService);
       }
 
       batteryService.setCharacteristic(
@@ -52,8 +53,8 @@ class Accessory {
         this.api.hap.Characteristic.ChargingState.NOT_CHARGEABLE
       );
     } else {
-      if (this.accessory.getService(this.api.hap.Service.BatteryService)) {
-        this.accessory.removeService(this.accessory.getService(this.api.hap.Service.BatteryService));
+      if (this.accessory.getService(this.BatteryService)) {
+        this.accessory.removeService(this.accessory.getService(this.BatteryService));
       }
     }