+++ src/accessories/smarthome/smarthome-button.handler.js
@@ -21,6 +21,7 @@ class Handler {
     this.accessories = accessories;
     this.polling = polling;
     this.fritzbox = meshMaster.fritzbox;
+    this.BatteryService = this.api.hap.Service.Battery || this.api.hap.Service.BatteryService;
 
     this.configured = true;
 
@@ -108,12 +109,12 @@ class Handler {
                   let lowBattery = device.battery.low || 0;
 
                   accessory
-                    .getService(this.api.hap.Service.BatteryService)
+                    .getService(this.BatteryService)
                     .getCharacteristic(this.api.hap.Characteristic.BatteryLevel)
                     .updateValue(batteryLevel);
 
                   accessory
-                    .getService(this.api.hap.Service.BatteryService)
+                    .getService(this.BatteryService)
                     .getCharacteristic(this.api.hap.Characteristic.StatusLowBattery)
                     .updateValue(lowBattery);
                 } else {
@@ -250,12 +251,12 @@ class Handler {
                   let lowBattery = device.battery.low || 0;
 
                   accessory
-                    .getService(this.api.hap.Service.BatteryService)
+                    .getService(this.BatteryService)
                     .getCharacteristic(this.api.hap.Characteristic.BatteryLevel)
                     .updateValue(batteryLevel);
 
                   accessory
-                    .getService(this.api.hap.Service.BatteryService)
+                    .getService(this.BatteryService)
                     .getCharacteristic(this.api.hap.Characteristic.StatusLowBattery)
                     .updateValue(lowBattery);
                 } else {