+++ src/accessories/smarthome/thermostat/thermostat.accessory.js
@@ -11,6 +11,7 @@ class Accessory {
     this.api = api;
     this.accessory = accessory;
     this.HistoryService = HistoryService;
+    this.BatteryService = this.api.hap.Service.Battery || this.api.hap.Service.BatteryService;
 
     this.handler = Handler.configure(api, accessories, accessory.context.config.polling, meshMaster);
     this.getService();
@@ -45,14 +46,14 @@ class Accessory {
     }
 
     if (this.accessory.context.config.battery) {
-      let batteryService = this.accessory.getService(this.api.hap.Service.BatteryService);
+      let batteryService = this.accessory.getService(this.BatteryService);
 
       if (!batteryService) {
         logger.info(
           'Adding Battery service',
           `${this.accessory.displayName} (${this.accessory.context.config.subtype})`
         );
-        batteryService = this.accessory.addService(this.api.hap.Service.BatteryService);
+        batteryService = this.accessory.addService(this.BatteryService);
       }
 
       batteryService.setCharacteristic(
@@ -60,8 +61,8 @@ class Accessory {
         this.api.hap.Characteristic.ChargingState.NOT_CHARGEABLE
       );
     } else {
-      if (this.accessory.getService(this.api.hap.Service.BatteryService)) {
-        this.accessory.removeService(this.accessory.getService(this.api.hap.Service.BatteryService));
+      if (this.accessory.getService(this.BatteryService)) {
+        this.accessory.removeService(this.accessory.getService(this.BatteryService));
       }
     }